+++ test/unit/vagrant/devcommands/internal_spec.rb
@@ -7,7 +7,7 @@ describe VagrantPlugins::DevCommands::Internal do
   describe 'displaying command list' do
     it 'includes internal commands' do
       cmd_foo      = command_def.new(name: 'foo', script: 'foo')
-      reg          = registry.new
+      reg          = registry.new(nil)
       reg.commands = { 'foo' => cmd_foo }
 
       internal = described_class.new(reg)
@@ -19,7 +19,8 @@ describe VagrantPlugins::DevCommands::Internal do
 
   describe 'internal help command' do
     it 'notifies if no command is available' do
-      internal = described_class.new(registry.new)
+      reg      = registry.new(nil)
+      internal = described_class.new(reg)
 
       expect { internal.run('help', []) }.to output(/no commands/i).to_stdout
     end
@@ -27,7 +28,7 @@ describe VagrantPlugins::DevCommands::Internal do
     it 'displays a list of available commands' do
       cmd_bar      = command_def.new(name: 'bar', script: 'bar')
       cmd_foo      = command_def.new(name: 'foo', script: 'foo')
-      reg          = registry.new
+      reg          = registry.new(nil)
       reg.commands = { 'bar' => cmd_bar, 'foo' => cmd_foo }
 
       internal = described_class.new(reg)
@@ -40,7 +41,7 @@ describe VagrantPlugins::DevCommands::Internal do
       cmd_bar      = command_def.new(name: 'bar', script: 'bar')
       cmd_foo      = command_def.new(name: 'foo', script: 'foo',
                                      desc: 'has a description')
-      reg          = registry.new
+      reg          = registry.new(nil)
       reg.commands = { 'bar' => cmd_bar, 'foo' => cmd_foo }
 
       internal = described_class.new(reg)
@@ -53,8 +54,9 @@ describe VagrantPlugins::DevCommands::Internal do
 
   describe 'internal version command' do
     it 'displays the plugin version' do
+      reg      = registry.new(nil)
+      internal = described_class.new(reg)
       version  = VagrantPlugins::DevCommands::VERSION
-      internal = described_class.new(registry.new)
 
       expect { internal.run('version', []) }.to output(/#{version}/).to_stdout
     end