+++ CHANGELOG.md
@@ -9,6 +9,8 @@
     the configured ui class (allows colored output)
   - Parameter wrapping is only done if a value is passed
     or a default (at least empty string) is configured
+  - Warnings about using internal command names or missing scripts are now
+    printed using the configured ui class (allows colored output)
 
 ## v0.5.0 (2015-03-21)